[HTMLTemplateElement] <template> inside of <head> may not create <body> if EOF is hit https://bugs.webkit.org/show_bug.cgi?id=109338
Reviewed by Adam Barth.

Source/WebCore:

This patch adds the logic to clear the stack of open elements back to the first <template> when EOF
is hit. This allows a <body> to be generated if the initial <template> was opened inside of <head>.

Tests added to html5lib.

* html/parser/HTMLTreeBuilder.cpp:
(WebCore):
(WebCore::HTMLTreeBuilder::popAllTemplates):
(WebCore::HTMLTreeBuilder::processEndTag):
(WebCore::HTMLTreeBuilder::processEndOfFile):
* html/parser/HTMLTreeBuilder.h:
(HTMLTreeBuilder):

LayoutTests:

* html5lib/resources/template.dat:


git-svn-id: svn://svn.chromium.org/blink/trunk@142635 bbb929c8-8fbe-4397-9dbb-9b2b20218538