LayoutUnit::epsilon shouldn't be necessary to place floats https://bugs.webkit.org/show_bug.cgi?id=94000
Reviewed by Levi Weintraub.

Currently, to get float placement correct, we need to have an
epsilon tolerance in RenderBlock::computeLogicalLocationForFloat
for the width of the line to fit them on. This really shouldn't
be necessary, and probably indicates we're losing precision
elsewhere.

Remove epsilon as it no longer appears to be needed. All layout
tests pass without it.

No new tests, covered by existing tests.

* rendering/RenderBlock.cpp:
(WebCore::RenderBlock::computeLogicalLocationForFloat):


git-svn-id: svn://svn.chromium.org/blink/trunk@143357 bbb929c8-8fbe-4397-9dbb-9b2b20218538