IndexedDB: Properly refactor frontend/backend code by #includes https://bugs.webkit.org/show_bug.cgi?id=110653
Reviewed by Dimitri Glazkov.

Source/WebCore:

IndexedDB has a distinct backend and frontend, with well defined
*Interface / *Callbacks classes distinguishing them, respectively.

This makes sure that the frontend and backend do not include each
other and fixes build problems that occurred as a result.

Specifically this moves enums out of IDBCursor and IDBTransaction
into IndexedDB.h in a new WebCore::IndexedDB namespace.

No new tests, as this is just a refactor.

Source/WebKit/chromium:

Clean up the frontend/backend WebIDB* implementations to
match the changes to webkit.


git-svn-id: svn://svn.chromium.org/blink/trunk@144820 bbb929c8-8fbe-4397-9dbb-9b2b20218538