Plumb a constraining ancestor node through moveParagraphs to createMarkup.
In InsertListCommand::unlistifyParagraph, it's important that the paragraph boundaries passed to
mergeParagraphs are respected, as otherwise the list item node isn't removed. For other cases,
we still want to preserve presentational nodes.

This is covered by the existing editing/pasteboard/paste-and-sanitize.html test, which was
naively rebaselined in http://trac.webkit.org/changeset/144995. This change reverts the
expectations from before that change.

Review URL: https://codereview.chromium.org/13867003

git-svn-id: svn://svn.chromium.org/blink/trunk@147970 bbb929c8-8fbe-4397-9dbb-9b2b20218538