Implement the DeviceAcceleration and DeviceRotationRate IDL objects
Remove the custom bindings --except for the event initialization
methods-- in favor of implementing the IDL objects as defined by
the specification. This patch adds the DeviceAcceleration and
DeviceRotationRate objects.

There should be no observable change in behavior. This is being
exercized by the existing layout tests.

BUG=WK112910 (already reviewed)

Review URL: https://codereview.chromium.org/13598014

git-svn-id: svn://svn.chromium.org/blink/trunk@148047 bbb929c8-8fbe-4397-9dbb-9b2b20218538