Can't use fastGetAttribute with WebCore::HTMLNames::classAttr when element is a SVGStyledElement.
On SVG elements, HTMLNames::classAttr is animatable so we can't always use the
fastGetAttribute. Fetching HTMLNames::classAttr is known to be a hot spot so
this change added a custom getClassAttribute (modeled on get(Id|Name)Attribute)
so we still use fastGetAttribute in 99% of the cases where it matters.

BUG=178030
R=abarth@chromium.org,eseidel@chromium.org

Review URL: https://chromiumcodereview.appspot.com/13872004

git-svn-id: svn://svn.chromium.org/blink/trunk@148322 bbb929c8-8fbe-4397-9dbb-9b2b20218538