Remove HTMLElement::adjustDirectionalityIfNeededAfterChildrenChanged
We can replace the complicated and broken adjustDirectionalityIfNeededAfterChildrenChanged
method with hooks on insertedInto and removedFrom to make HTMLElement's implementation
of the HTML5 dir=auto feature simpler. This old method had unreachable code that checked
childCountDelta < 0 and then had a loop with count = 0; count < childCountDelta; count++
so the loop never actually did anything.

After this patch we can replace childCountDelta with an enum with values ChildrenAdded and
ChildrenRemoved. There's no reason to keep track of the actual number as nothing actually
uses that value.

Review URL: https://codereview.chromium.org/14159008

git-svn-id: svn://svn.chromium.org/blink/trunk@148399 bbb929c8-8fbe-4397-9dbb-9b2b20218538