Add temporary double-to-float range truncation to MediaSourcePrivateImpl::duration().
Currently this range truncation logic lives in Chromium code
(WebMediaSourceClientImpl::duration()). It needs to be moved
into MediaDourcePrivateImpl::duration() so that when the
Chromium code is updated to properly return durations in the
double range, tests in Blink won't start failing. This code
will be removed in a follow-up change that converts the
WebMediaPlayer xxxFloat() signatures to their double
versions.

BUG=227156
TEST=Existing LayoutTests still pass. No functional change since this truncation is currently
     happening in Chromium code as well.

Review URL: https://codereview.chromium.org/13925022

git-svn-id: svn://svn.chromium.org/blink/trunk@148412 bbb929c8-8fbe-4397-9dbb-9b2b20218538