Remove left-over PageCache stuff
PageCache was removed in r148973. This patch removes the rest
of the Settings code (except a stub which is waiting on a Chromium
patch to take care of it) as well as various comments/logging code.

Finally, it gets rid of LayoutTests that are no longer exercising
anything interesting. In a few cases, the test still seems to be
testing interesting behavior; these have been updated to remove
their reference to the UsesPageCache pref.

R=abarth
BUG=229605

Review URL: https://codereview.chromium.org/14481003

git-svn-id: svn://svn.chromium.org/blink/trunk@149044 bbb929c8-8fbe-4397-9dbb-9b2b20218538